Boy was this annoying to debug.
If you call BWindow::AddShortcut() with a message parameter but not a target,
the target becomes the window. In this case WebWindow does not handle B_UNDO or
B_REDO.  Adding the NULL parameter as the target in AddShortcut forces the
shortcut handling to find the window's current focus and send the message
there. This is how all the other text editing shortcuts (B_SELECT_ALL, B_CUT,
etc) are set up in BWindow which is why they worked but Undo/Redo didn't.

Given all the other code related to undo/redo works I'm surprised this was
broken. Maybe Stephan didn't test after the change which introduced these here.

Long story short, undo/redo now work when editing in Web+.

Fixes at least #6224.

git-svn-id: http://svn.haiku-os.org/webpositive/webkit/trunk@583 94f232f2-1747-11df-bad5-a5bfde151594