Merge 108780 - [v8] when a named item on document goes out of scope, actually remove it
https://bugs.webkit.org/show_bug.cgi?id=79409

Reviewed by Adam Barth.

Source/WebCore:

The original change already included the code, but it led to some
problems, so it was reverted in http://trac.webkit.org/changeset/63845.
However, not removing the items leaks a considerable amount of memory,
so I'm adding the code back. The bug that led to the revert was that
the accessor was removed unconditionally. I'm now only removing it,
when the last item with that name is removed.

Test: fast/dom/HTMLDocument/named-item.html

* bindings/v8/V8DOMWindowShell.cpp:
(WebCore::V8DOMWindowShell::namedItemRemoved):

LayoutTests:

* fast/dom/HTMLDocument/named-item-expected.txt: Added.
* fast/dom/HTMLDocument/named-item.html: Added.

TBR=jochen@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9667038