Bring back (incomplete) html5 canvas
As the StillImage is now created in ImageBuffer's constructor (before drawing),
it needs to be able to work on a shared BBitmap instead of a copy. cf. Qt port.
Until recently, the StillImage copy was created in ImageBuffer::image(), after
drawing happened.

For the same reason, we also need to Sync() the BView that draws in the
BBitmap when we return or draw the image.

This brings back html5 canvas support to at least the level it was before the
port upgrade (if not better). This means that it still misses basic stuff like
gradient and transforms and has rendering bugs.

Noticed a regression in curve drawing.