Fix various rounded rectangle issues
There are two code paths for drawing the background of rounded rects, depending
on the border width (1px or >1px).
Be sure to use alpha blending when drawing the rounded rect background via
clipping and also when drawing the border arcs.
Fix two errors in the math when drawing the background of rounded rects via
bezier arc approximation (it seems that this code path wasn't really used
before).

Rounded rects should work better than before (according to my testing) although
it's still not perfect in some edge cases. The current drawing code path is
deprecated in upstream webkit anyway, and will be ditched when we support
polygon and shape clipping.